Add enrollment history to guardians swimling panel.
