+++ swimling_dashboard/templates/swimling_dashboard/_public_lessons_panel.html
@@ -35,12 +35,19 @@
                   </h4>
                 </div>
               </div>
-              <button onclick="openEditSwimlingModal({{ swimling.id }})"
-                 data-edit-url="{% url 'swimling_dashboard:edit-swimling' id=swimling.id %}"
-                class="inline-flex items-center px-4 py-2 bg-white border border-gray-300 text-gray-600 hover:bg-gray-50 hover:border-gray-400 rounded-lg transition-all duration-200 shadow-sm hover:shadow-md text-sm font-medium"
-                title="Edit {{swimling.first_name}}'s details">
-                <i class="fa-solid fa-user-pen mr-2"></i>Edit Details
-              </button>
+              <div class="flex items-center gap-2">
+                <button onclick="openEditSwimlingModal({{ swimling.id }})"
+                   data-edit-url="{% url 'swimling_dashboard:edit-swimling' id=swimling.id %}"
+                  class="inline-flex items-center px-4 py-2 bg-white border border-gray-300 text-gray-600 hover:bg-gray-50 hover:border-gray-400 rounded-lg transition-all duration-200 shadow-sm hover:shadow-md text-sm font-medium"
+                  title="Edit {{swimling.first_name}}'s details">
+                  <i class="fa-solid fa-user-pen mr-2"></i>Edit Details
+                </button>
+                <button type="button"
+                  class="inline-flex items-center px-4 py-2 bg-purple-600 text-white text-sm font-medium rounded-lg shadow-sm hover:bg-purple-700 transition disabled:opacity-50 disabled:cursor-not-allowed"
+                  {% if swimling.previous_terms %}onclick="openHistoryModal('dashboard-history-{{ swimling.id }}')"{% else %}disabled{% endif %}>
+                  <i class="fas fa-history mr-2"></i>Enrollment History
+                </button>
+              </div>
             </div>
           </div>
 
@@ -167,11 +174,19 @@
                 </div>
                 <span class="font-bold text-gray-900 text-lg">{{ swimling.first_name }} {{ swimling.last_name }}</span>
               </div>
-              <button onclick="openEditSwimlingModal({{ swimling.id }})"
-                 data-edit-url="{% url 'swimling_dashboard:edit-swimling' id=swimling.id %}"
-                class="text-gray-500 hover:text-gray-700 p-2" title="Edit {{swimling.first_name}}'s details">
-                <i class="fa-solid fa-user-pen text-lg"></i>
-              </button>
+              <div class="flex items-center gap-2">
+                <button onclick="openEditSwimlingModal({{ swimling.id }})"
+                   data-edit-url="{% url 'swimling_dashboard:edit-swimling' id=swimling.id %}"
+                  class="text-gray-500 hover:text-gray-700 p-2" title="Edit {{swimling.first_name}}'s details">
+                  <i class="fa-solid fa-user-pen text-lg"></i>
+                </button>
+                <button type="button"
+                  class="text-purple-600 hover:text-purple-700 p-2 disabled:text-gray-400"
+                  {% if swimling.previous_terms %}onclick="openHistoryModal('dashboard-history-{{ swimling.id }}')"{% else %}disabled{% endif %}
+                  title="View previous enrollments">
+                  <i class="fas fa-history text-lg"></i>
+                </button>
+              </div>
             </div>
           </div>
 
@@ -270,9 +285,40 @@
             {% endif %}
           </div>
         </div>
-        {% endwith %}
-        {% endfor %}
+{% endwith %}
+{% endfor %}
       </div>
+
+      {% for item in public_lessons_data %}
+      {% with swimling=item.swimling %}
+      <div id="dashboard-history-{{ swimling.id }}" class="fixed inset-0 z-50 hidden bg-black bg-opacity-50 flex items-center justify-center p-4">
+        <div class="bg-white rounded-xl shadow-2xl w-full max-w-lg">
+          <div class="flex items-center justify-between px-4 py-3 border-b border-gray-200">
+            <h3 class="text-base font-semibold text-gray-900">Enrollment History</h3>
+            <button type="button" class="text-gray-500 hover:text-gray-700"
+              onclick="closeHistoryModal('dashboard-history-{{ swimling.id }}')">
+              <i class="fas fa-times"></i>
+            </button>
+          </div>
+          <div class="px-4 py-3 max-h-[70vh] overflow-y-auto">
+            {% if swimling.previous_terms %}
+            <div class="space-y-3">
+              {% for entry in swimling.previous_terms %}
+              <div class="border border-gray-200 rounded-lg p-3 bg-gray-50">
+                <p class="text-gray-900 text-sm font-semibold">{{ entry.level }}</p>
+                <p class="text-gray-600 text-xs">{{ entry.day }}{% if entry.time %} • {{ entry.time }}{% endif %}</p>
+                <p class="text-gray-500 text-xs">{{ entry.term }}</p>
+              </div>
+              {% endfor %}
+            </div>
+            {% else %}
+            <p class="text-sm text-gray-500">No previous enrollments recorded.</p>
+            {% endif %}
+          </div>
+        </div>
+      </div>
+      {% endwith %}
+      {% endfor %}
     {% else %}
       <div class="text-center py-12">
         <div class="w-16 h-16 bg-gray-100 rounded-full flex items-center justify-center mx-auto mb-4">
